Output percentiles, sample/cor matrices to file again instead of json
As it turns out, cytoscape table rows are assumed to reference networks,
nodes or edges. Not all this extra data fits that assumption, e.g.
percentiles. Even cor matrices reference genes, which is not a 1-to-1
mapping to nodes. It's much easier and makes more sense to just dump to
file instead of trying to load them into cytoscape; due to that
assumption.